[FIX] Rename generic bitstream.h to cc_bitstream.h #1436 (#1543)
* fix: rename bitstream

* fix: update vcpkg commit hash

* fix: try to fix the linker error
