+++ .github/workflows/build_windows.yml
@@ -40,8 +40,8 @@ jobs:
         id: vcpkg
         uses: friendlyanon/setup-vcpkg@v1
         with:
-          committish: "217194cea7e3491c14779c763beec50b6a4575d6"
-          cache-version: "1"
+          committish: "2023.06.20"
+          cache-version: "2"
           ignore-reserve-cache-error: true
       - name: Install dependencies
         run: cd vcpkg && vcpkg integrate install && vcpkg install leptonica tesseract ffmpeg --triplet x64-windows-static
@@ -76,8 +76,8 @@ jobs:
         id: vcpkg
         uses: friendlyanon/setup-vcpkg@v1
         with:
-          committish: "217194cea7e3491c14779c763beec50b6a4575d6"
-          cache-version: "1"
+          committish: "2023.06.20"
+          cache-version: "2"
           ignore-reserve-cache-error: true
       - name: Install dependencies
         run: cd vcpkg && vcpkg integrate install && vcpkg install leptonica tesseract ffmpeg --triplet x64-windows-static
